2 - Software Engineering (SE)/Week 15/2 - Tutorial/multi-container-app/app/requirements.txt
Flask
mysql-connector-python
